# LibJS: Speed up IndexedPropertyIterator by computing non-empty indices